Working on GUI tabs. Currently only placeholders. Removed filename header info.
