+++ lib/config.lua
@@ -1,6 +1,3 @@
--- config.lua
--- Aug 2024
-
 --[[
   __  __  ___  ___   __      ___   ___ _  _ ___ _  _  ___
  |  \/  |/ _ \|   \  \ \    / /_\ | _ \ \| |_ _| \| |/ __|
