+++ lib/gui_tabs/regrowth_controls.lua
@@ -1,0 +1,11 @@
+-- Contains the GUI for the regrowth controls tab.
+
+-- Used by AddOarcGuiTab
+function CreateRegrowthControlsTab(tab_container, player)
+    AddLabel(tab_container, nil, "test regrowth message my_label_style", my_label_style)
+    AddLabel(tab_container, nil, "test regrowth message my_label_header_style", my_label_header_style)
+    AddLabel(tab_container, nil, "test regrowth message my_label_header_grey_style", my_label_header_grey_style)
+    AddLabel(tab_container, nil, "test regrowth message my_note_style", my_note_style)
+    AddLabel(tab_container, nil, "test regrowth message my_warning_style", my_warning_style)
+    AddLabel(tab_container, nil, "test regrowth message my_longer_label_style", my_longer_label_style)
+end
